requirements.txt
fastapi==0.128.0
openai==2.14.0
passlib==1.7.4
pydantic==2.12.5
pydantic_settings==2.12.0
python_docx==1.2.0
python_jose==3.5.0
Requests==2.32.5
streamlit==1.52.2
supabase==2.27.0
uvicorn==0.40.0
bcrypt==4.1.2
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
# httpx==0.25.2
httpx==0.27.0
pytest-mock==3.12.0
email-validator==2.1.0 
python-multipart
pydantic[email]